RN-635: Broke MeditrakSyncQueue into separate classes - MeditrakSyncQueue acts as the ChangeHandler - MeditrakSyncRecordUpdater ensures the sync queue records are updated when new changes come in
