+++ packages/central-server/src/database/index.js
@@ -3,7 +3,6 @@
  * Copyright (c) 2017 Beyond Essential Systems Pty Ltd
  */
 
-export { MeditrakSyncQueue } from './MeditrakSyncQueue';
-export { createPermissionsBasedMeditrakSyncQueue } from './createPermissionsBasedMeditrakSyncQueue';
+export { MeditrakSyncQueue, createPermissionsBasedMeditrakSyncQueue } from './meditrakSyncQueue';
 export * from './utilities';
 export * from './constants';
