+++ packages/central-server/src/database/meditrakSyncQueue/index.js
@@ -1,0 +1,7 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2022 Beyond Essential Systems Pty Ltd
+ */
+
+export { MeditrakSyncQueue } from './MeditrakSyncQueue';
+export { createPermissionsBasedMeditrakSyncQueue } from './createPermissionsBasedMeditrakSyncQueue';
